+++ packages/babel-preset-react-app/create.js
@@ -6,7 +6,17 @@
  */
 'use strict';
 
-const { validateBoolOption } = require('./utils');
+const validateBoolOption = (name, value, defaultValue) => {
+  if (typeof value === 'undefined') {
+    value = defaultValue;
+  }
+
+  if (typeof value !== 'boolean') {
+    throw new Error(`Preset react-app: '${name}' option must be a boolean.`);
+  }
+
+  return value;
+};
 
 module.exports = function(api, opts, env) {
   if (!opts) {
