--- packages/babel-preset-react-app/utils.js
@@ -1,21 +1,0 @@
-/**
- * Copyright (c) 2015-present, Facebook, Inc.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-'use strict';
-
-const validateBoolOption = (name, value, defaultValue) => {
-  if (typeof value === 'undefined') {
-    value = defaultValue;
-  }
-
-  if (typeof value !== 'boolean') {
-    throw new Error(`Preset react-app: '${name}' option must be a boolean.`);
-  }
-
-  return value;
-};
-
-module.exports = { validateBoolOption };
